# Conflicts:
#	code/modules/power/gravitygenerator.dm

+++ code/_onclick/hud/screen_objects.dm
@@ -28,7 +28,10 @@
 	screen_loc = null
 	if(length(hud?.mymob?.client?.screen))
 		hud.mymob.client.screen -= src
-	hud = null
+	// All register signals MUST be mirrored with Unregister signals included in a Destroy proc, even if other procs also call Unregister Signal.
+	if (hud)
+		UnregisterSignal(hud, COMSIG_QDELETING)
+		hud = null
 	return ..()
 
 /// Screen elements are always on top of the players screen and don't move so yes they are adjacent
@@ -40,6 +43,10 @@
  */
 /atom/movable/screen/proc/handle_hud_destruction()
 	SIGNAL_HANDLER
+
+	// If we were QDEL'ed directly (probably by our owner), then there's no need to Unregister Signal and qdel ourselves twice.
+	if (QDELING(src))
+		return
 
 	UnregisterSignal(hud, COMSIG_QDELETING)
 	qdel(src)
